Mostramos todas las llaves para estar listos a certificar...
git-svn-id: http://192.168.1.94/svn@601 d43fff5d-296d-41c8-bd8c-7388abecf751

+++ siyen_2010/trunk/modules/siyen-core/src/main/java/edu/ienpop/service/impl/CursoSinCertificarServiceImpl.java
@@ -29,6 +29,9 @@ public class CursoSinCertificarServiceImpl implements CursoSinCertificarService 
 	public void borraCursoSinCertificar(Long idCurso) {
 		CursoSinCertificar cursoSinCertificar = cursoSinCertificarDao
 				.read(idCurso);
+		if(cursoSinCertificar.isListoParaCertificar()){
+			throw new BusinessException("No se puede borrar este curso por que ya esta listo para certificarse...");
+		}
 		cursoSinCertificarDao.delete(cursoSinCertificar);
 	}
 
@@ -56,7 +59,7 @@ public class CursoSinCertificarServiceImpl implements CursoSinCertificarService 
 
 	public CursoSinCertificar obtenerCursoSinCertificarConAlumnos(Long idCurso) {
 		CursoSinCertificar cursoSinCertificar = cursoSinCertificarDao
-				.read(idCurso);
+				.obtenerCursoSinCertificarConRelaciones(idCurso);
 		List<AlumnoSinCertificar> alumnosSinCertificar = alumnoSinCertificaroDao
 				.obtenerAlumnosSinCertificarPorIdCurso(idCurso);
 		cursoSinCertificar.setAlumnosSinCertificar(alumnosSinCertificar);
